+++ CFQ/Utils/Enum/LogLevel.swift
@@ -1,0 +1,9 @@
+
+import Foundation
+
+enum LogLevel: String {
+    case info = "ℹ️ INFO"
+    case action = "💥 ACTION"
+    case warning = "⚠️ WARNING"
+    case error = "❌ ERROR"
+}
